% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDBN.R
\name{HDBN}
\alias{HDBN}
\title{R6 abstract class that defines the classifier + DBN model}
\arguments{
\item{lower}{vector of lower bounds of the variables to optimize}

\item{upper}{vector of upper bounds of the variables to optimize}

\item{itermax}{maximum number of iterations of the optimization process}

\item{test_per}{percentage of instances assigned as test in the optimization}

\item{optim_trace}{whether or not to print the progress of each optimization iteration}

\item{id_var}{an index variable that identifies different time series in the data}

\item{cl_obj_var}{the objective variable for the classifier}

\item{dbn_obj_vars}{the objective variables for the DBN}

\item{horizon}{integer value that defines the lenght of the forecasting with the DBN model}

\item{dt_test}{a data.table with the test dataset}

\item{print_res}{a boolean that determines whether or not should the results of the prediction be printed}

\item{conf_mat}{a boolean that determines whether or not should a confusion matrix be printed}

\item{size}{the size of the DBN}

\item{method}{the structure learning method used}

\item{...}{additional parameters for the DBN structure learning}

\item{dt_train}{a data.table with the training dataset}

\item{optim}{boolean that determines whether or not the classifier parameters should be optimized}

\item{cl_params}{the classifier parameters.}

\item{dt}{a data.table with the training dataset}
}
\value{
the prediction result vector

the prediction result vector
}
\description{
Initialize the object with some modifiable parameters of the optimization

Fit the classifier and the DBN models to some provided data

Predict the objective variable in all the rows in a dataset with the 
classifier augmented by the DBN forecasting. The horizon sets the length of 
the forecasting with the DBN model.

Predict the objective variable in all the rows in a dataset using only
the classifier model.

Print the parameters used to train the model

Fit the internal DBN

Fit the internal classifier

Optimize the parameters of the internal classifier
}
\details{
The model is defined as a DBN model that forecasts the state of a time-series
based system and a classifier that classifies that predicted state.
This is a very simplified abstract class for all the different hybrid models
that I would like to try. It could be more generic, but I do not have the
time to encapsulate it any further.
}
\section{Fields}{

\describe{
\item{\code{cl}}{the classifier model}

\item{\code{cl_params}}{the classifier parameters.}

\item{\code{cl_obj_var}}{the variable to be classified with the classifier}

\item{\code{net}}{the DBN graph object}

\item{\code{fit}}{the DBN fitted model}

\item{\code{dbn_obj_vars_raw}}{the variables to be predicted with the DBN model without the appended 't_0'}

\item{\code{dbn_obj_vars}}{the variables to be predicted with the DBN model}

\item{\code{id_var}}{the variable that differentiates each independent individual TS in the dataset}

\item{\code{size}}{size argument of the DBN}

\item{\code{optim_lower}}{lower bounds of the parameters}

\item{\code{optim_upper}}{bounds of the parameters}

\item{\code{optim_itermax}}{maximum number of iterations of the optimization process}

\item{\code{optim_test_per}}{percentage of instances assigned as test in the optimization}

\item{\code{optim_trace}}{whether or not to print the progress of each optimization iteration}
}}

