% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGDBN.R
\name{XGDBN}
\alias{XGDBN}
\title{R6 class that defines the XGBoost + DBN model}
\arguments{
\item{lower}{lower bounds of the weight, max tree depth and number of rounds}

\item{upper}{upper bounds of the weight, max tree depth and number of rounds}

\item{itermax}{maximum number of iterations of the optimization process}

\item{test_per}{percentage of instances assigned as test in the optimization}

\item{optim_trace}{whether or not to print the progress of each optimization iteration}

\item{id_var}{an index variable that identifies different time series in the data}

\item{size}{the size of the DBN}

\item{method}{the structure learning method used}

\item{cl_obj_var}{the objective variable for the XGBoost}

\item{dbn_obj_vars}{the objective variables for the DBN}

\item{...}{additional parameters for the DBN structure learning}

\item{horizon}{integer value that defines the lenght of the forecasting with the DBN model}

\item{dt_test}{a data.table with the test dataset}

\item{print_res}{a boolean that determines whether or not should the results of the prediction be printed}

\item{conf_mat}{a boolean that determines whether or not should a confusion matrix be printed}

\item{dt_train}{a data.table with the training dataset}

\item{optim}{boolean that determines wheter or not the classifier parameters should be optimized}

\item{cl_params}{the classifier parameters.}
}
\value{
the prediction result vector

the prediction result vector
}
\description{
Initialize the object with some modifiable parameters of the optimization

Fit the XGBoost and the DBN models to some provided data

Predict the objective variable in all the rows in a dataset with the 
XGBoost augmented by the DBN forecasting. The horizon sets the length of 
the forecasting with the DBN model.

Predict the objective variable in all the rows in a dataset using only
the XGBoost model.

Fit the internal XGBoost
}
\details{
The model is defined as a DBN model that forecasts the state of a time-series
based system and an XGBoost model that classifies that predicted state.
}
